[dsutil] gzip reader rejects non-gzip files
for safety

+++ accelerator/standard_methods/a_csvimport.py
@@ -191,14 +191,16 @@ def prepare(job, slices):
 		# re-use import logic
 		out_fns = ["labels"]
 		r_num = cstuff.mk_uint64(3)
-		open("labels", "wb").close()
 		try:
 			import_slice("c backend failed in label parsing", labels_rfd, -1, -1, -1, out_fns, b"wb1", separator, r_num, quote_char, lf_char, 0, 0)
 		finally:
 			os.close(labels_rfd)
-		with typed_reader("bytes")("labels") as fh:
-			labels_from_file = [lab.decode("utf-8", "backslashreplace") for lab in fh]
-		os.unlink("labels")
+		if os.path.exists("labels"):
+			with typed_reader("bytes")("labels") as fh:
+				labels_from_file = [lab.decode("utf-8", "backslashreplace") for lab in fh]
+			os.unlink("labels")
+		else:
+			labels_from_file = []
 	else:
 		labels_from_file = None
